refactor: BooleanGenerator와 구현체 util 패키지로 이동
