+++ src/main/java/ladder/domain/Line.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import ladder.util.BooleanGenerator;
 
 public class Line {
