+++ src/test/java/ladder/util/TestBooleanGenerator.java
@@ -1,7 +1,6 @@
 package ladder.util;
 
 import java.util.List;
-import ladder.domain.BooleanGenerator;
 
 public class TestBooleanGenerator implements BooleanGenerator {
